+++ iOS/MyNumberCardAuth/MyNumberCardAuthTests/Controllers/AuthenticationControllerTests.swift
@@ -8,31 +8,7 @@ import XCTest
 @testable import MyNumberCardAuth
 import SwiftUI
 
-public class AuthenticationManagerMock: AuthenticationManagerProtocol {
-    var barCallCount = 0
-    var pin = ""
-    var nonce = ""
-    var actionURL = ""
-    
-    public func authenticateForSignature(pin: String, nonce: String, actionURL: String, authenticationController:AuthenticationController) {
-        self.pin = pin
-        self.nonce = nonce
-        self.actionURL = actionURL
-        barCallCount += 1
-    }
-    
-    public func authenticateForUserVerification(pin: String, nonce: String, actionURL: String, authenticationController:AuthenticationController) {
-        self.pin = pin
-        self.nonce = nonce
-        self.actionURL = actionURL
-        barCallCount += 1
-    }
-}
-
 final class AuthenticationControllerTests: XCTestCase {
-
-    let controller:AuthenticationController = AuthenticationController()
-    
     override func setUpWithError() throws {
         // Put setup code here. This method is called before the invocation of each test method in the class.
     }
@@ -41,7 +17,9 @@ final class AuthenticationControllerTests: XCTestCase {
         // Put teardown code here. This method is called after the invocation of each test method in the class.
     }
 
-    func testclear() throws {
+    func testClear() throws {
+        let controller:AuthenticationController = AuthenticationController()
+        
         controller.isAlert = true
         controller.isLinkAlert = true
         controller.isErrorOpenURL = true
@@ -57,40 +35,46 @@ final class AuthenticationControllerTests: XCTestCase {
     }
     
     func testOpenURLButton() throws {
+        let controller:AuthenticationController = AuthenticationController()
+        
         // mock必要
         controller.openURLButton(url:"https://example.com")
     }
     
     func testOpenURLButtonURLEmpty() throws {
+        let controller:AuthenticationController = AuthenticationController()
+        
         controller.openURLButton(url:"")
     }
     
     func testOpenURLButtonNoURL() throws {
+        let controller:AuthenticationController = AuthenticationController()
+        
         controller.openURLButton(url:"てすと")
     }
 
-    func testStartReading()throws {
+    func testStartReading() throws {
         let mock = AuthenticationManagerMock()
-        let authenticationController = AuthenticationController(authenticationManager: mock)
-            
-        authenticationController.viewState = .SignatureView
-        authenticationController.startReading(pin: "1234", nonce: "0123456789", actionURL: "https:example.1")
+        let controller = AuthenticationController(authenticationManager: mock)
+        
+        controller.viewState = .SignatureView
+        controller.startReading(pin: "1234", nonce: "0123456789", actionURL: "https:example.1")
         
         XCTAssertEqual(mock.barCallCount,1)
         XCTAssertEqual(mock.pin,"1234")
         XCTAssertEqual(mock.nonce,"0123456789")
         XCTAssertEqual(mock.actionURL,"https:example.1")
         
-        authenticationController.viewState = .UserVerificationView
-        authenticationController.startReading(pin: "5678", nonce: "9876543210", actionURL: "https:example.2")
+        controller.viewState = .UserVerificationView
+        controller.startReading(pin: "5678", nonce: "9876543210", actionURL: "https:example.2")
         
         XCTAssertEqual(mock.barCallCount,2)
         XCTAssertEqual(mock.pin,"5678")
         XCTAssertEqual(mock.nonce,"9876543210")
         XCTAssertEqual(mock.actionURL,"https:example.2")
         
-        authenticationController.viewState = .ExplanationView
-        authenticationController.startReading(pin: "test", nonce: "nonce", actionURL: "https:example.3")
+        controller.viewState = .ExplanationView
+        controller.startReading(pin: "test", nonce: "nonce", actionURL: "https:example.3")
         
         XCTAssertEqual(mock.barCallCount,2)
         XCTAssertEqual(mock.pin,"5678")
@@ -99,29 +83,35 @@ final class AuthenticationControllerTests: XCTestCase {
     }
     
     func testGetButtonColor() throws {
+        let controller:AuthenticationController = AuthenticationController()
         XCTAssertEqual(controller.getButtonColor(checkStr: "example"), Color(UIColor.blue))
     }
     
     func testGetButtonColorEmpty() throws {
+        let controller:AuthenticationController = AuthenticationController()
         XCTAssertEqual(controller.getButtonColor(checkStr: ""), Color(UIColor.lightGray))
     }
     
     func testSetErrorPageURLQueryDicNil() throws {
+        let controller:AuthenticationController = AuthenticationController()
         controller.setErrorPageURL(queryDict: ["":""])
         XCTAssertEqual(controller.openURL, "")
     }
     
     func testSetErrorPageURLNoErrorURL() throws {
+        let controller:AuthenticationController = AuthenticationController()
         controller.setErrorPageURL(queryDict: ["error_url":""])
         XCTAssertEqual(controller.openURL, "")
     }
     
     func testSetErrorPageURL() throws {
+        let controller:AuthenticationController = AuthenticationController()
         controller.setErrorPageURL(queryDict: ["error_url":"https://example/?test=1&amp;test2=1"])
         XCTAssertEqual(controller.openURL, "https://example/?test=1&test2=1")
     }
     
     func testOnOpenURLEmptyNonceEmpty() throws {
+        let controller:AuthenticationController = AuthenticationController()
         controller.onOpenURL(url: URL(string: "http://example?action_url=example")!)
         XCTAssertEqual(controller.controllerForUserVerification.actionURL,"")
         XCTAssertEqual(controller.controllerForUserVerification.nonce,"")
@@ -130,6 +120,7 @@ final class AuthenticationControllerTests: XCTestCase {
     }
     
     func testOnOpenURLEmptyActionURLEmpty() throws {
+        let controller:AuthenticationController = AuthenticationController()
         controller.onOpenURL(url: URL(string: "http://example?nonce=example")!)
         XCTAssertEqual(controller.controllerForUserVerification.actionURL,"")
         XCTAssertEqual(controller.controllerForUserVerification.nonce,"")
@@ -138,6 +129,7 @@ final class AuthenticationControllerTests: XCTestCase {
     }
     
     func testOnOpenURL() throws {
+        let controller:AuthenticationController = AuthenticationController()
         controller.onOpenURL(url: URL(string: "https://example?action_url=example&nonce=1234")!)
         XCTAssertEqual(controller.controllerForUserVerification.actionURL,"example")
         XCTAssertEqual(controller.controllerForUserVerification.nonce,"1234")
