+++ iOS/MyNumberCardAuth/MyNumberCardAuthTests/Controllers/SignatureViewControllerTests.swift
@@ -9,9 +9,7 @@ import XCTest
 import SwiftUI
 
 final class SignatureViewControllerTests: XCTestCase {
-    
-    let controller:SignatureViewController = SignatureViewController()
-    
+        
     override func setUpWithError() throws {
         // Put setup code here. This method is called before the invocation of each test method in the class.
     }
@@ -21,6 +19,8 @@ final class SignatureViewControllerTests: XCTestCase {
     }
     
     func testIsEnableButton() throws {
+        let controller:SignatureViewController = SignatureViewController()
+        
         controller.inputPIN = ""
         controller.nonce = ""
         controller.actionURL = ""
@@ -48,6 +48,8 @@ final class SignatureViewControllerTests: XCTestCase {
     }
     
     func testGetButtonColor() throws {
+        let controller:SignatureViewController = SignatureViewController()
+        
         controller.inputPIN = ""
         controller.nonce = ""
         controller.actionURL = ""
