+++ iOS/MyNumberCardAuth/MyNumberCardAuthTests/Controllers/UserVerificationViewControllerTests.swift
@@ -9,9 +9,7 @@ import XCTest
 import SwiftUI
 
 final class UserVerificationViewControllerTests: XCTestCase {
-    
-    let controller:UserVerificationViewController = UserVerificationViewController()
-    
+        
     override func setUpWithError() throws {
         // Put setup code here. This method is called before the invocation of each test method in the class.
     }
@@ -21,6 +19,8 @@ final class UserVerificationViewControllerTests: XCTestCase {
     }
     
     func testIsEnableButton() throws {
+        let controller:UserVerificationViewController = UserVerificationViewController()
+        
         controller.inputPIN = ""
         controller.nonce = ""
         controller.actionURL = ""
@@ -53,6 +53,8 @@ final class UserVerificationViewControllerTests: XCTestCase {
     }
     
     func testGetButtonColor() throws {
+        let controller:UserVerificationViewController = UserVerificationViewController()
+        
         controller.inputPIN = ""
         controller.nonce = ""
         controller.actionURL = ""
