+++ iOS/MyNumberCardAuth/MyNumberCardAuthTests/Models/AuthenticationManagerTests.swift
@@ -18,7 +18,9 @@ final class AuthenticationManagerTests: XCTestCase {
         // Put teardown code here. This method is called after the invocation of each test method in the class.
     }
     
-    
-    
-    
+    func testAuthenticateForUserVerification() throws {
+
+
+
+    }
 }
