+++ iOS/MyNumberCardAuth/MyNumberCardAuthTests/MyNumberCardAuthAppTests.swift
@@ -10,8 +10,6 @@ import XCTest
 
 final class MyNumberCardAuthAppTests: XCTestCase {
 
-    let myNumberApp = MyNumberCardAuthApp()
-    
     override func setUpWithError() throws {
         // Put setup code here. This method is called before the invocation of each test method in the class.
     }
