+++ iOS/MyNumberCardAuth/MyNumberCardAuthUITests/MyNumberCardAuthUITests.swift
@@ -7,6 +7,7 @@
 
 import XCTest
 
+// ユニバーサルリンクを使用するため、keycloakへのアクセスが可能な状態にすること
 var keycloakURL: String = "https://852ed96220eb.ngrok.app" //ポート80
 var relum: String = "OIdp"
